Fix all problems with map displaying and some uint32 -> uint64 Use getUInt64 for GUID. Use old system for world_edit (because the new system is unstable, boost::multi_array<Tile, 2> is hard to use) And minor fixes.

+++ server/Player.hpp
@@ -31,7 +31,7 @@ class WOD_DLL_DECL Player : public Unit
     friend class WorldSession;
 
 public:
-    Player(std::string Username, std::string Password, uint32 ObjID);
+    Player(std::string Username, std::string Password, uint64 ObjID);
     ~Player();
 
     // TODO
